% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_to_renviron}
\alias{add_to_renviron}
\title{Add a variable to a user's ~/.Renviron file}
\usage{
add_to_renviron(x)
}
\arguments{
\item{x}{The text that you'd like to add, e.g. \code{VARIABLE=value}}
}
\value{
\code{TRUE} (in)
}
\description{
Add a variable to a user's ~/.Renviron file
}

