% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_push_pull.R
\name{cache_version_files}
\alias{cache_version_files}
\title{A function for caching information about document state}
\usage{
cache_version_files(doc_id, source, rendered_md,
  cache_dir = getOption("gd.cache"))
}
\arguments{
\item{doc_id}{The Google ID of the document}

\item{source}{The Rmd source file at the time of the run}

\item{rendered_md}{The rendered md of the Rmd source file at the time of the
run}

\item{cache_dir}{The dir used for caching}
}
\value{
Nothing
}
\description{
A function for caching information about document state
}

