% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R, R/misc.R
\name{file_path}
\alias{file_path}
\title{Construct Path to File}
\usage{
file_path(..., fsep = .Platform$file.sep)

file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors}

\item{fsep}{the path separator to use.}

\item{...}{character vectors}

\item{fsep}{the path separator to use.}
}
\value{
A character vector

A character vector
}
\description{
A modified version of \code{\link{file.path}}, which prevents \code{fsep}
cropping up more than once.

A modified version of \code{\link{file.path}}, which prevents \code{fsep}
cropping up more than once.
}

