% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_auth}
\alias{gd_auth}
\title{Authorize with Google Documents}
\usage{
gd_auth(client_id, client_secret, cache = "~/.googdown_token")
}
\arguments{
\item{client_id}{Unimplemented}

\item{client_secret}{Unimplemented}

\item{cache}{passed to httr}
}
\value{
Used for it's side effects.
}
\description{
Authorize with Google Documents
}

