% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{gd_delete}
\alias{gd_delete}
\title{Delete a file in a Google Drive account}
\usage{
gd_delete(file_id)
}
\arguments{
\item{file_id}{The id of the file / document to delete}
}
\value{
The content of the httr response. Used for its side effects.
}
\description{
Delete a file in a Google Drive account
}
\author{
br
}

