% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_download}
\alias{gd_download}
\title{Download a Google Document as Markdown or JSON}
\usage{
gd_download(doc_id, file_name = tempfile(),
  export_format = defaultDownloadFormat(), output_format = "markdown",
  revision = NA)
}
\arguments{
\item{doc_id}{The ID of the Google document}

\item{file_name}{File path to download to}

\item{export_format}{The format the document is exported from google using}

\item{output_format}{The format for the final outputted file}
}
\value{
If successful, \code{file_name}
}
\description{
A small wrapper function for \code{\link{gd_export}}, which converts the
exported file to markdown or JSON using pandoc
}

