% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_export}
\alias{gd_export}
\title{Download an A Google Doc as an Open Office or MS Word Document}
\usage{
gd_export(doc_id, file_name, export_format = getOption("gd.download_format"),
  revision = NA)
}
\arguments{
\item{doc_id}{The ID of the Google document}

\item{file_name}{File path to download to}

\item{export_format}{The format the document is exported from google using}

\item{revision}{The revision id which you'd like to download the document
at. A value of \code{NA} (the default) downloads the latest version}
}
\value{
The httr response
}
\description{
Download an A Google Doc as an Open Office or MS Word Document
}

