% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{gd_ls}
\alias{gd_ls}
\title{Get a list of the files in a g Google Drive account}
\usage{
gd_ls()
}
\value{
The content of the httr response
}
\description{
Get a list of the files in a g Google Drive account
}

