% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_push_pull.R
\name{gd_pull}
\alias{gd_pull}
\title{Pull remote changes to a Google Doc to the source Rmarkdown file}
\usage{
gd_pull(file_name, format = defaultUploadFormat())
}
\arguments{
\item{file}{The path to the Rmarkdown file which you'd like to update}
}
\value{
\code{TRUE} (invisibly) if successfull, otherwise, an error.
}
\description{
Pull remote changes to a Google Doc to the source Rmarkdown file
}

