% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{gd_update}
\alias{gd_update}
\title{Update the contents of a Google Doc, with a local file}
\usage{
gd_update(local_file, doc_id)
}
\arguments{
\item{local_file}{The file which you'd like to upload}

\item{doc_id}{the google doc id}
}
\value{
Used for its side effects.
}
\description{
Update the contents of a Google Doc, with a local file
}

