% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_download.R
\name{gd_upload}
\alias{gd_upload}
\title{Upload a file to Google Docs}
\usage{
gd_upload(file_name, format = defaultUploadFormat())
}
\arguments{
\item{file_name}{The name of the local file which you'd like to upload}

\item{format}{The conversion format you'd like to use to upload it}
}
\value{
The httr response
}
\description{
Upload a file to Google Docs
}

