% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{google_doc}
\alias{google_doc}
\title{Convert a Local Rmarkdown File to a remote Google Doc}
\usage{
google_doc(reference_file = NULL,
  upload_format = getOption("gd.upload_format"), keep_md = FALSE,
  clean_supporting = FALSE)
}
\arguments{
\item{reference_file}{The reference file to be used for styling. The default
(\code{NULL}) is to use a template with the same styling as a default
Google doc, though with slightly nicer figure captions.}

\item{upload_format}{The format the Rmarkdown document should be converted
to, prior to upload}

\item{keep_md}{Should the intermediate markdown files be retained?}

\item{clean_supporting}{Should other intermediate files be retained?}
}
\value{
If successful, the URL of a remote Google document
}
\description{
This function is intended to be used as an \code{rmarkdown} 'output format',
e.g. \code{rmarkdown::render("myfile.Rmd", output_format =
googdown::google_doc())}. You can also get the same result via
\code{\link{gd_push}}.
}

