% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_push_pull.R
\name{post_processor}
\alias{post_processor}
\title{googdown's post processor (run after pandoc)}
\usage{
post_processor(metadata, input_file, output_file, clean, verbose)
}
\arguments{
\item{metadata}{Supplied by \code{rmarkdown::\link{output_format}}}

\item{input_file}{Supplied by \code{rmarkdown::\link{output_format}}}

\item{output_file}{Supplied by \code{rmarkdown::\link{output_format}}}

\item{clean}{Supplied by \code{rmarkdown::\link{output_format}}}

\item{verbose}{Supplied by \code{rmarkdown::\link{output_format}}}
}
\value{
Used for its side effects
}
\description{
Uploads the file to google docs, adds the doc_id to the YAML front matter,
and caches some local files.
}
\keyword{internal}

