% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_push_pull.R
\name{pre_knit}
\alias{pre_knit}
\title{googdown pre-knit hook}
\usage{
pre_knit(input)
}
\arguments{
\item{input}{The file path of the Rmarkdown source code}
}
\value{
Used for its side effects
}
\description{
It simply checks google auths, and then 'standardizes' the Rmarkdown file,
using \code{\link{standardize_rmd}}
}
\keyword{internal}

