% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_alter.R
\name{remove_docx_bookmarks}
\alias{remove_docx_bookmarks}
\title{Remove 'Bookmark References' from a .docx file}
\usage{
remove_docx_bookmarks(input_file)
}
\arguments{
\item{input_file}{The .docx file you'd like the bookmark references stripped
from}
}
\value{
TRUE (invisibly) if it worked
}
\description{
'Bookmark References' are MS Word's internal version of header I.D.s, for
linking to or navigating documents. Pandoc adds Bookmark References to all
.docx output to try and be helpful. Unfortunately, Google Docs renders these
with big blue ugly ribbons by the side of each header. This function strips
them out of a given input file.
}
\details{
MS Word files are zipped XML. The file is unzipped, and a regular expression
is used to parse and remove the offending bookmarks from the XML. Then
everything is zipped back together again.
}

