% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_conversion.R
\name{standardize_rmd}
\alias{standardize_rmd}
\title{Convert a file to commonmark, using system pandoc}
\usage{
standardize_rmd(input_file)
}
\arguments{
\item{input_file}{The file you'd like to conver}
}
\value{
Used for its side effects
}
\description{
Convert a file to commonmark, using system pandoc
}

