% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_download.R
\name{gd_download}
\alias{gd_download}
\title{Download an A Google Doc as Markdown}
\usage{
gd_download(file_id, file_name = "./file.md",
  format = defaultDownloadFormat())
}
\arguments{
\item{file_id}{The Google Doc ID of the document you'd like to download}

\item{file_name}{The local file you'd like the Markdown data stored in}

\item{format}{The conversion format you'd like to use}
}
\value{
The httr response
}
\description{
Download an A Google Doc as Markdown
}

