% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_export}
\alias{gd_export}
\title{Download an A Google Doc as an Open Office or MS Word Document}
\usage{
gd_export(doc_id, file_name, export_format = getOption("gd.download_format"),
  revision = NA)
}
\arguments{
\item{doc_id}{The Google Doc ID of the document you'd like to download}

\item{file_name}{The local file you'd like to export to}

\item{format}{The export format you'd like to use}
}
\value{
The httr response
}
\description{
Download an A Google Doc as an Open Office or MS Word Document
}

