% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_push_pull.R
\name{gd_push}
\alias{gd_push}
\title{Push an Rmarkdown document to Google Docs}
\usage{
gd_push(file_name)
}
\arguments{
\item{file_name}{The filepath of the Rmarkdown source code}
}
\value{
\code{TRUE} (invisibly) if successfull, otherwise, an error.
}
\description{
Push an Rmarkdown document to Google Docs
}

