% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{gd_push}
\alias{gd_push}
\title{Push an Rmarkdown document to Google Docs}
\usage{
gd_push(file_name, upload_format = getOption("gd.upload_format"),
  reference_file = NULL)
}
\arguments{
\item{file_name}{An Rmarkdown file}

\item{upload_format}{The format the Rmarkdown document should be converted
to, prior to upload}

\item{reference_file}{The reference file to be used for styling. The default
(\code{NULL}) is to use a template with the same styling as a default
Google doc, though with slightly nicer figure captions.}
}
\value{
\code{TRUE} (invisibly) if successfull, otherwise, an error.
}
\description{
Push an Rmarkdown document to Google Docs
}

