% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{gd_push}
\alias{gd_push}
\title{Push an Rmarkdown document to Google Docs}
\usage{
gd_push(file_name, bookdown_md = TRUE, reference_docx = NULL)
}
\arguments{
\item{file_name}{An Rmarkdown file}

\item{bookdown_md}{Should markdown extensions by bookdown by used?}

\item{reference_docx}{The reference file to be used for styling. The default
(\code{NULL}) is to use a template with the same styling as a default
Google doc, though with slightly nicer figure captions.}
}
\value{
\code{TRUE} (invisibly) if successfull, otherwise, an error.
}
\description{
Push an Rmarkdown document to Google Docs
}

