% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{gdoc_push}
\alias{gdoc_push}
\title{Push an Rmarkdown document to Google Docs}
\usage{
gdoc_push(file_name, format = defaultUploadFormat(), reference_odt = NULL)
}
\arguments{
\item{file_name}{The file path of the Rmarkdown source code}
}
\value{
\code{TRUE} (invisibly) if successfull, otherwise, an error.
}
\description{
Push an Rmarkdown document to Google Docs
}

